refactor(solver): split subtype/core.rs under the 2000-line cap
Move the ~1.8k-line structural dispatch method `check_subtype_inner_impl`
out of `subtype/core.rs` into a new child module `subtype/core_dispatch.rs`
(registered via `#[path] mod` so it inherits the parent's scope through
`use super::*`). The method is bumped to `pub(crate)` so its sole caller
(the `stacker::maybe_grow` closure in `check_subtype_inner`) can reach it
across the module boundary.

Behavior-preserving relocation: `core.rs` drops from 2854 to 995 lines and
`core_dispatch.rs` is 1876, both under the §19 hard 2000-line cap, clearing
the Reviewer file-size blocker on PR #10626. Full `tsz-solver` lib suite
shows no new failures vs the pre-existing baseline; clippy clean.
